# docs/reference/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "DeliveryBundleDescriptor": {
        "description": "Structured delivery descriptor shared across frontend replay/export paths.",
        "properties": {
          "filenameBase": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Filenamebase",
            "type": "string"
          },
          "htmlDocumentTitle": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Htmldocumenttitle",
            "type": "string"
          },
          "htmlSections": {
            "items": {
              "$ref": "#/components/schemas/DeliveryBundleSection"
            },
            "maxItems": 12,
            "title": "Htmlsections",
            "type": "array"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/DeliveryBundleMetric"
            },
            "maxItems": 16,
            "title": "Metrics",
            "type": "array"
          },
          "shareContent": {
            "maxLength": 50000,
            "minLength": 1,
            "title": "Sharecontent",
            "type": "string"
          },
          "subagentTrail": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Subagenttrail",
            "type": "array"
          },
          "summary": {
            "default": "",
            "maxLength": 5000,
            "title": "Summary",
            "type": "string"
          },
          "summaryLines": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Summarylines",
            "type": "array"
          },
          "title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "title",
          "filenameBase",
          "shareContent",
          "htmlDocumentTitle"
        ],
        "title": "DeliveryBundleDescriptor",
        "type": "object"
      },
      "DeliveryBundleMetric": {
        "description": "One delivery overview metric surfaced to share/export clients.",
        "properties": {
          "label": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "tone": {
            "anyOf": [
              {
                "enum": [
                  "default",
                  "success",
                  "warning",
                  "danger",
                  "info"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tone"
          },
          "value": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "label",
          "value"
        ],
        "title": "DeliveryBundleMetric",
        "type": "object"
      },
      "DeliveryBundlePayload": {
        "description": "Full delivery bundle persisted behind share links.",
        "properties": {
          "artifact": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact"
          },
          "descriptor": {
            "$ref": "#/components/schemas/DeliveryBundleDescriptor"
          },
          "executionReceipt": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executionreceipt"
          },
          "htmlContent": {
            "maxLength": 200000,
            "minLength": 1,
            "title": "Htmlcontent",
            "type": "string"
          },
          "schemaVersion": {
            "const": "2026-03-29",
            "default": "2026-03-29",
            "title": "Schemaversion",
            "type": "string"
          },
          "share": {
            "$ref": "#/components/schemas/DeliveryBundleShareMetadata"
          }
        },
        "required": [
          "descriptor",
          "htmlContent",
          "share"
        ],
        "title": "DeliveryBundlePayload",
        "type": "object"
      },
      "DeliveryBundleSection": {
        "description": "Structured section used by delivery HTML and replay snapshots.",
        "properties": {
          "items": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Items",
            "type": "array"
          },
          "key": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Key",
            "type": "string"
          },
          "title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "key",
          "title"
        ],
        "title": "DeliveryBundleSection",
        "type": "object"
      },
      "DeliveryBundleShareMetadata": {
        "description": "Share metadata co-packaged with delivery artifacts.",
        "properties": {
          "content": {
            "maxLength": 50000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ],
        "title": "DeliveryBundleShareMetadata",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ShareCreateRequest": {
        "additionalProperties": false,
        "description": "Create share-link request body.",
        "properties": {
          "content": {
            "maxLength": 50000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "delivery_bundle": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeliveryBundlePayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "html_content": {
            "anyOf": [
              {
                "maxLength": 200000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Html Content"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "content"
        ],
        "title": "ShareCreateRequest",
        "type": "object"
      },
      "ShareCreateResponse": {
        "description": "Create share-link response body.",
        "properties": {
          "share_id": {
            "title": "Share Id",
            "type": "string"
          },
          "share_url": {
            "title": "Share Url",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "share_id",
          "share_url"
        ],
        "title": "ShareCreateResponse",
        "type": "object"
      },
      "ShareDetailResponse": {
        "description": "Shared content response body.",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "delivery_bundle": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeliveryBundlePayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "html_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Html Content"
          },
          "share_id": {
            "title": "Share Id",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "share_id",
          "content",
          "created_at"
        ],
        "title": "ShareDetailResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI Travel Assistant API with SSE streaming support.",
    "title": "moyuan-travel-agent API",
    "version": "3.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/share-links": {
      "post": {
        "description": "Create a short share id and return an app URL with share query parameter.",
        "operationId": "create_share_link_api_share_links_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Share Link",
        "tags": [
          "share"
        ]
      }
    },
    "/api/share-links/{share_id}": {
      "get": {
        "description": "Fetch shared travel-plan content by share id.",
        "operationId": "get_share_link_api_share_links__share_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "share_id",
            "required": true,
            "schema": {
              "maxLength": 10,
              "minLength": 10,
              "pattern": "^[a-f0-9]{10}$",
              "title": "Share Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Share Link",
        "tags": [
          "share"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Snapshot server",
      "url": "http://localhost:38083"
    }
  ]
}